move default impl of genesis into tests
